Refactor and run in tests
