+++ docs/examples/frameworks/jax/index.rst
@@ -7,4 +7,5 @@ JAX Framework
    jax-basic_example.ipynb
    jax-multigpu_example.ipynb
    flax-basic_example.ipynb
+   pax-basic_example.ipynb
