+++ docs/examples/frameworks/jax/pax_examples/dali_pax_input.py
