+++ qa/TL0_jupyter/test_jax.sh
@@ -1,11 +1,12 @@
 #!/bin/bash -e
 # used pip packages
-pip_packages='jupyter numpy jax flax'
+pip_packages='jupyter numpy jax flax paxml'
 target_dir=./docs/examples
 
 test_body() {
     test_files=(
         "frameworks/jax/jax-basic_example.ipynb"
+        "frameworks/jax/pax-basic_example.ipynb"
     )
     for f in ${test_files[@]}; do
         jupyter nbconvert --to notebook --inplace --execute \
